% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sGM12878.R
\docType{data}
\name{sGM12878}
\alias{sGM12878}
\title{Promoter Capture data for chromosomes 20 and 21 in GM12878}
\format{A ChicagoData object.}
\source{
Mifsud, B. et al. "Mapping long-range promoter contacts in human cells with high-resolution capture Hi-C." Nature Genetics (2015) doi:10.1038/ng.3286
}
\usage{
sGM12878
}
\description{
Promoter Capture data to be used as a toy example to run all steps of Chicago. This data only incorporates read pairs
including both chromosomes 20 and 21 in order to minimize processing time and memory storage. Thus, this data set
includes all cis read pairs for these two chromosomes and all trans read pairs between them.
}
\examples{
data(sGM12878)
##modifications to sGM12878, ensuring it uses correct design directory
designDir <- file.path(system.file("extdata", package="PCHiCdata"), "hg19TestDesign")
sGM12878 <- modifySettings(cd = sGM12878, designDir=designDir)
}
\seealso{
\code{\link{chicagoData}}
}
\keyword{datasets}

